arm, riscv: shift unsigned values, not promoted signed ones
Six shifts had undefined results. On ARM the RRX path shifted a signed
int into its sign bit, and the 32-bit ROM, flash and SRAM reads compose
a word as byte | byte<<8 | byte<<16 | byte<<24, where each uint8_t is
promoted to int and the last shift is undefined for any byte >= 0x80.
That is the memory path every ARM workload uses constantly. On RISC-V the
JAL, branch and store immediates were sign-extended by shifting a signed
value left and back.

Both now shift through uint32_t. The emulators are built -O3 -flto and
depend on bit-exact results, so this is worth closing even though no
miscompilation has been observed: UBSan is quiet across the ARM and
RISC-V workloads afterwards, arm-decode and arm-jit are unchanged, and
arm-bench is unchanged.

+++ src/arm/arm_cpu.c
@@ -193,18 +193,18 @@ static inline arm_io_region_t *io_lookup(arm_cpu_t *cpu, uint32_t *addr,
 uint32_t arm_read32(arm_cpu_t *cpu, uint32_t addr) {
     addr &= ~3u;
     if (cpu->rom && addr < cpu->rom_size) {
-        return cpu->rom[addr] | (cpu->rom[addr+1]<<8) |
-               (cpu->rom[addr+2]<<16) | (cpu->rom[addr+3]<<24);
+        return (uint32_t)cpu->rom[addr] | ((uint32_t)cpu->rom[addr+1]<<8) |
+               ((uint32_t)cpu->rom[addr+2]<<16) | ((uint32_t)cpu->rom[addr+3]<<24);
     }
     if (addr >= cpu->flash_base && addr < cpu->flash_end) {
         uint32_t off = addr - cpu->flash_base;
-        return cpu->flash[off] | (cpu->flash[off+1]<<8) |
-               (cpu->flash[off+2]<<16) | (cpu->flash[off+3]<<24);
+        return (uint32_t)cpu->flash[off] | ((uint32_t)cpu->flash[off+1]<<8) |
+               ((uint32_t)cpu->flash[off+2]<<16) | ((uint32_t)cpu->flash[off+3]<<24);
     }
     if (addr >= cpu->sram_base && addr < cpu->sram_end) {
         uint32_t off = addr - cpu->sram_base;
-        return cpu->sram[off] | (cpu->sram[off+1]<<8) |
-               (cpu->sram[off+2]<<16) | (cpu->sram[off+3]<<24);
+        return (uint32_t)cpu->sram[off] | ((uint32_t)cpu->sram[off+1]<<8) |
+               ((uint32_t)cpu->sram[off+2]<<16) | ((uint32_t)cpu->sram[off+3]<<24);
     }
     /* Bit-band alias for peripheral region */
     if (addr >= ARM_BITBAND_BASE && addr < ARM_BITBAND_BASE + 0x02000000) {
@@ -247,13 +247,13 @@ static inline uint32_t mem_read32(arm_cpu_t *cpu, uint32_t addr) {
     if (arm_tz_blocks(cpu, addr)) return 0;
     if (__builtin_expect(addr >= cpu->sram_base && addr < cpu->sram_end, 1)) {
         uint32_t off = addr - cpu->sram_base;
-        return cpu->sram[off] | (cpu->sram[off+1]<<8) |
-               (cpu->sram[off+2]<<16) | (cpu->sram[off+3]<<24);
+        return (uint32_t)cpu->sram[off] | ((uint32_t)cpu->sram[off+1]<<8) |
+               ((uint32_t)cpu->sram[off+2]<<16) | ((uint32_t)cpu->sram[off+3]<<24);
     }
     if (addr >= cpu->flash_base && addr < cpu->flash_end) {
         uint32_t off = addr - cpu->flash_base;
-        return cpu->flash[off] | (cpu->flash[off+1]<<8) |
-               (cpu->flash[off+2]<<16) | (cpu->flash[off+3]<<24);
+        return (uint32_t)cpu->flash[off] | ((uint32_t)cpu->flash[off+1]<<8) |
+               ((uint32_t)cpu->flash[off+2]<<16) | ((uint32_t)cpu->flash[off+3]<<24);
     }
     return arm_read32(cpu, addr);
 }
@@ -3088,7 +3088,7 @@ static int arm_step_interpreter(arm_cpu_t *cpu, int count) {
                                 /* RRX */
                                 int old_c = (cpu->xpsr & APSR_C) ? 1 : 0;
                                 carry_out = rm_val & 1;
-                                rm_val = (rm_val >> 1) | (old_c << 31);
+                                rm_val = (rm_val >> 1) | ((uint32_t)old_c << 31);
                             } else {
                                 rm_val = ror_c(rm_val, shift_n, &carry_out);
                             }
